+++ arquillian-steps/src/main/java/io/fabric8/kubernetes/pipeline/arquillian/cube/kubernetes/InNamespaceStep.java
@@ -22,31 +22,69 @@ import org.jenkinsci.plugins.workflow.steps.StepExecution;
 import org.kohsuke.stapler.DataBoundConstructor;
 
 import java.io.Serializable;
-import java.util.List;
 import java.util.Map;
 
 import hudson.Extension;
 
-public class SessionStep extends AbstractSessionManagerStep implements Serializable {
+public class InNamespaceStep extends AbstractStep implements Serializable {
 
     private static final long serialVersionUID = 5588861066775717487L;
 
+    private final String name;
+    private final String prefix;
+
+    private final Map<String, String> labels;
+    private final Map<String, String> annotations;
+
+    private final Boolean namespaceLazyCreateEnabled;
+    private final Boolean namespaceDestroyEnabled;
+
     @DataBoundConstructor
-    public SessionStep(String cloud, String name, String prefix, Map<String, String> labels, Map<String, String> annotations, String environmentSetupScriptUrl, String environmentTeardownScriptUrl, String environmentConfigUrl, List<String> environmentDependencies, Long waitTimeout, List<String> waitForServiceList, Boolean namespaceLazyCreateEnabled, Boolean namespaceCleanupEnabled, Boolean namespaceDestroyEnabled) {
-       super(cloud, name, prefix, labels, annotations, environmentSetupScriptUrl, environmentTeardownScriptUrl, environmentConfigUrl, environmentDependencies, waitTimeout, waitForServiceList, namespaceLazyCreateEnabled, namespaceCleanupEnabled, namespaceDestroyEnabled);
+    public InNamespaceStep(String cloud, String name, String prefix, Map<String, String> labels, Map<String, String> annotations, Boolean namespaceLazyCreateEnabled, Boolean namespaceDestroyEnabled) {
+        super(cloud);
+        this.name = name;
+        this.prefix = prefix;
+        this.labels = labels;
+        this.annotations = annotations;
+
+        this.namespaceLazyCreateEnabled = namespaceLazyCreateEnabled != null ? namespaceLazyCreateEnabled : true;
+        this.namespaceDestroyEnabled = namespaceDestroyEnabled != null ? namespaceDestroyEnabled : true;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getPrefix() {
+        return prefix;
+    }
+
+    public Map<String, String> getLabels() {
+        return labels;
+    }
+
+    public Map<String, String> getAnnotations() {
+        return annotations;
+    }
+
+    public Boolean isNamespaceLazyCreateEnabled() {
+        return namespaceLazyCreateEnabled;
+    }
+
+    public Boolean isNamespaceDestroyEnabled() {
+        return namespaceDestroyEnabled;
     }
 
     @Override
     public StepExecution start(StepContext context) throws Exception {
-        return new SessionStepExecution(this, context);
+        return new InNamespaceStepExecution(this, context);
     }
-
 
     @Extension
     public static class DescriptorImpl extends AbstractStepDescriptorImpl {
 
        public DescriptorImpl() {
-            super(SessionStepExecution.class);
+            super(InNamespaceStepExecution.class);
         }
 
         public DescriptorImpl(Class<? extends StepExecution> executionType) {
@@ -55,12 +93,12 @@ public class SessionStep extends AbstractSessionManagerStep implements Serializa
 
         @Override
         public String getFunctionName() {
-            return "inSession";
+            return "inNamespace";
         }
 
         @Override
         public String getDisplayName() {
-            return "Run build steps inside an Arquillian Cube Kubernetes Session";
+            return "Run build steps inside an arquillian cube kubernetes managed namespace";
         }
 
         @Override
